Create README.md
In this code, I use the same idea as described in Section 3.2 in the paper of Vargas et. al. There are some versions go with this. In this version, I train one Neural Network for the control of the drift to minimize the original loss (i.e. the loss without any further computation and simplification). Moreover, in the paper of Vargas et. al., we need to define a prescribe a curve of distribution p_t. This can be chosen as the interpolation between the reference distribution p_0 = N(mu, sigma) and the target pi. However, as can be seen in the paper of Bálint Máté and François Fleuret (see https://arxiv.org/pdf/2301.07388), the naive linear interpolation can cause some training problem due the mass teleportation in some cases. Thus, using the idea in the paper of Bálint Máté and François Fleuret, in this code, I also train the interpolation path between p_0 and pi. Finally, I also try to train the prior parameters mu and sigma_ref of p_0.